# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'E-commerce Modular API'
  description: 'Tài liệu API hệ thống E-commerce (Sử dụng PHP 8 Attributes)'
  contact:
    email: admin@system.com
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API Server'
paths:
...
  /admin/promotions:
    get:
      tags:
        - Promotions
      summary: 'Lấy danh sách khuyến mãi'
      operationId: 815501f4796ccd641aaf926b1f9bdaa4
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
        -
          name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Promotions
      summary: 'Tạo mới khuyến mãi'
      operationId: 8280d34b6047f93dd7aefad48126a922
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - type
                - value
                - start_date
                - end_date
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum:
                    - percentage
                    - fixed
                value:
                  type: number
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                product_ids:
                  type: array
                  items:
                    type: integer
              type: object
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/admin/promotions/{uuid}':
    get:
      tags:
        - Promotions
      summary: 'Xem chi tiết khuyến mãi'
      operationId: 90f53a56a3e227f3c48e96cfebcc29ed
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Promotions
      summary: 'Cập nhật khuyến mãi'
      operationId: bad7f0f192a9b0095bc1390f390d90a6
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                status:
                  type: boolean
              type: object
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Promotions
      summary: 'Xóa khuyến mãi'
      operationId: 2103693e1ba3059dba187ed87a4db80a
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
  /reviews:
    get:
      tags:
        - Reviews
      summary: 'Xem danh sách đánh giá'
      operationId: 537151cfc85509ee58521c13903a8fe3
      parameters:
        -
          name: product_uuid
          in: query
          schema:
            type: string
            format: uuid
        -
          name: page
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success
    post:
      tags:
        - Reviews
      summary: 'Tạo đánh giá mới'
      operationId: fb71a3b9781fdda4cb83976b10b69418
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_uuid
                - rating
                - content
              properties:
                product_uuid:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  maximum: 5
                  minimum: 1
                content:
                  type: string
              type: object
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/reviews/{uuid}':
    put:
      tags:
        - Reviews
      summary: 'Cập nhật đánh giá'
      operationId: a15331cbca679ed871fedeae4de9f987
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              properties:
                rating:
                  type: integer
                content:
                  type: string
              type: object
      responses:
        200:
          description: Updated
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Reviews
      summary: 'Xóa đánh giá'
      operationId: 4620691994824d42f0925ef738ba349f
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []